Refactor notifications and Filters
